# go/model_login_data.go
/*
 * PingCastle API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: v1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type LoginData struct {

	Apikey string `json:"apikey,omitempty"`

	Location string `json:"location,omitempty"`
}