# go/model_report_data.go
/*
 * PingCastle API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: v1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ReportData struct {

	Filename string `json:"filename,omitempty"`

	XmlReport string `json:"xmlReport,omitempty"`
}